연락처 - 사용자 초대 화면 진행중

+++ lib/features/authentication/views/user_reg_req_screen.dart
@@ -116,7 +116,7 @@ class UserRegisterRequestScreenState extends ConsumerState<UserRegisterRequestSc
     );
   }
 
-  Widget _getRegIndex(double index) {
+  Widget _getRegIndex(int index) {
     if (index == 1) {
       return const BelongAgency(
         titles: [
